user.ex: Updated User schema to include the password and the hashed password fields. Also added changeset functions and a function to add the hash to the changeset (since it is a derived field).
mix.exs: Added Pbkdf2 dependency for hashing.

accounts.ex: Added functions to the Accounts context that would tie a changeset to a form and also functions that serve as a public facing API for our controllers to use. That way the controllers only interact with the context modules, and the rest is abstracted out to user.ex.

router.ex: Utilized the resource routes for index, show, new, and create.

user_controller.ex: Added the create function to register new users by interacting with the Accounts context.

new.html.eex: Template for the registration form that uses a changeset in conjunction with the form. It's a nifty Phoenix trick that makes creating forms pretty simple.