user_controller.ex: created an authentication function that redirects you to the home page if you are not logged in and there is no user in the session. We turned this into a function plug by giving it the required opts parameter then we specified at the top that the authenticate plug would always fire before the index and show actions executed. The authenticate function first checks if there is a current user in 'assigns'. The module plug is what puts the user there.
auth.ex: specifies the module plug that just checks if the session user exists and stores the user in the session as the current user. Also it contains the login function that initially stores the user in the session. login is called in the create action.

router.ex: we added the auth module plug to router.ex in the browser pipeline. the browser pipeline is called everytime for every route after /