Used the mix phx.gen.schema to generate a Category schema and migration to create the Categories table.
We also altered the Video schema to include a referential constraint to the Categories table. We make sure to cast the category_id in the changeset so our API can safely use it. This is different from what we did with the User reference in the Video schema because with the User reference, we do not expect the User to have to input that information, but the User will have to input the Category information.

We also created a blank migration to alter the Videos table in the database to include the category_id foreign_id.